+++ example_test.go
@@ -6,6 +6,10 @@ func ExamplePage_WriteTo() {
 	article := Article(
 		Class("fancy"),
 		H1("Title of my article"),
+		P(
+			"before ", A(Href("http://example.com"), "example"),
+			" after",
+		),
 	)
 	page := NewPage("", Html(Body(article)))
 	page.WriteTo(os.Stdout)
@@ -17,7 +21,7 @@ func ExamplePage_WriteTo() {
 	// <body>
 	// <article class="fancy">
 	// <h1>Title of my article</h1>
-	//
+	// <p>before <a href="http://example.com">example</a> after</p>
 	// </article>
 	// </body>
 	// </html>