+++ changelog.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.8.3] 2020-04-10
+### Added
+
+- Em tag
+- Attributes action, autocomplete, method, tabindex, value, fieldset
+
 ### Changed
 
 - htmlwriter adds newlines after table and tr tags